idcparse: keep track of line_being_executed
needed it to debug what some issue was, it was with enum member
serial values. When creating two enum member with the same value,
they will get serial 0 and 1. Deleting the first member leaves one
member for that value with serial 1. idcparse can't handle that
because it always assumes serials are sequential starting from 0.

Solution is to remove the other enum member and re-add it, so ida
will assign it serial 0 again.