fix docgen segfault wrt chained struct refs
if a ref was chaining struct members but a member's type was not
defined (as opposed to defined but not found), it would segfault.
Oops.

Also update the line in blogx3 that made me discover this (note that
it didn't error anymore since last commit because I updated the type
there already).