docgen: fix a resolving reference case
When the address is not the base offset of a struct member but it's
less than the offset of the next member, the resolving process would
go deeper (if the member is a struct itself) and continue.

It didn't work when this struct member was the last member of the
struct, due to a simple logic flaw that wasn't thought of before.